requirements.txt
# ---------------------------------------------------------------
# Core dependencies
# ---------------------------------------------------------------
pydantic>=2.7.0
pydantic-settings>=2.3.0

# Document processing
PyMuPDF>=1.24.0
python-docx>=1.1.0
docx2pdf>=0.1.8
Pillow>=10.3.0
openpyxl>=3.1.2
extract-msg>=0.48.0
html2image>=2.0.4
striprtf>=0.0.26

# Image processing
opencv-python-headless>=4.9.0.80
numpy>=1.26.0

# Layout detection
ultralytics>=8.2.0

# HTTP client
httpx>=0.27.0

# Async
asyncio-throttle>=1.0.2

# Queue
celery[redis]>=5.4.0
redis>=5.0.4

# Observability
structlog>=24.2.0
prometheus-client>=0.20.0
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp-proto-grpc>=1.24.0

# Storage
motor>=3.4.0
asyncpg>=0.29.0
alembic>=1.13.1
sqlalchemy>=2.0.30
neo4j>=5.20.0
qdrant-client>=1.9.0
pgvector>=0.3.0

# Utilities
pyyaml>=6.0.1
python-magic>=0.4.27
